+++ core/templates/dev/head/pages/exploration_player/exploration_player.html
@@ -42,7 +42,7 @@
 {% endblock navbar_breadcrumb %}
 
 {% block local_top_nav_options %}
-  {% include 'pages/exploration_player/learner_local_nav.html' %}
+  <learner-local-nav></learner-local-nav>
 {% endblock local_top_nav_options %}
 
 {% block content %}
