+++ core/templates/dev/head/pages/exploration_player/learner_local_nav_directive.html
@@ -1,25 +1,24 @@
-<ul class="nav ng-cloak navbar-nav oppia-navbar-nav navbar-right"
-    ng-controller="LearnerLocalNav" style="margin-right: 0px;">
-  <li popover-placement="bottom" uib-popover-template="'<[getFeedbackPopoverUrl()]>'" popover-trigger="click">
+<ul class="nav ng-cloak navbar-nav oppia-navbar-nav navbar-right" style="margin-right: 0px;">
+  <li popover-placement="bottom" uib-popover-template="'<[$ctrl.getFeedbackPopoverUrl()]>'" popover-trigger="click">
     <a href="" class="protractor-test-exploration-feedback-popup-link" uib-tooltip="<['I18N_PLAYER_FEEDBACK_TOOLTIP' | translate]>" tooltip-placement="bottom">
       <i class="material-icons">&#xE87F;</i>
       <span class="oppia-icon-accessibility-label"><['I18N_PLAYER_FEEDBACK_TOOLTIP' | translate]></span>
     </a>
   </li>
   <li>
-    <a ng-if="canEdit" ng-href="/create/<[explorationId]>" uib-tooltip="<['I18N_PLAYER_EDIT_TOOLTIP' | translate]>" tooltip-placement="bottom" target="_blank" rel="noopener">
+    <a ng-if="$ctrl.canEdit" ng-href="/create/<[$ctrl.explorationId]>" uib-tooltip="<['I18N_PLAYER_EDIT_TOOLTIP' | translate]>" tooltip-placement="bottom" target="_blank" rel="noopener">
       <i class="material-icons">&#xE254;</i>
       <span class="oppia-icon-accessibility-label"><['I18N_PLAYER_EDIT_TOOLTIP' | translate]></span>
     </a>
   </li>
   <li>
-    <a ng-if="!canEdit && username" class="protractor-test-exploration-suggestion-popup-link" ng-click="showLearnerSuggestionModal()" uib-tooltip="Suggest Changes" tooltip-placement="bottom" target="_blank" rel="noopener">
+    <a ng-if="!$ctrl.canEdit && $ctrl.username" class="protractor-test-exploration-suggestion-popup-link" ng-click="$ctrl.showLearnerSuggestionModal()" uib-tooltip="Suggest Changes" tooltip-placement="bottom" target="_blank" rel="noopener">
       <i class="material-icons">&#xE254;</i>
       <span class="oppia-icon-accessibility-label">Suggest Changes</span>
     </a>
   </li>
   <li>
-    <a ng-if="username" ng-click="showFlagExplorationModal()" uib-tooltip="<['I18N_PLAYER_REPORT_TOOLTIP' | translate]>" tooltip-placement="bottom" tabindex="0">
+    <a ng-if="$ctrl.username" ng-click="$ctrl.showFlagExplorationModal()" uib-tooltip="<['I18N_PLAYER_REPORT_TOOLTIP' | translate]>" tooltip-placement="bottom" tabindex="0">
       <i class="material-icons">&#xE153;</i>
       <span class="oppia-icon-accessibility-label"><['I18N_PLAYER_REPORT_TOOLTIP' | translate]></span>
     </a>
